+++ src/Service/S3/CHANGELOG.md
@@ -12,6 +12,7 @@
 - Added `S3Client::putPublicAccessBlock()` method
 - AWS api-change: Amazon Simple Storage Service / Features: Add conditional writes in CopyObject on destination key to prevent unintended object modifications.
 - AWS api-change: Adds support for blocking SSE-C writes to general purpose buckets.
+- AWS api-change: New S3 Storage Class FSX_ONTAP
 
 ### Dependency bumped